feat: Get ERC20 Transfer widget working again.
