+++ src/components/Guilds/ActionsModal/index.tsx
@@ -1,6 +1,8 @@
+import { utils } from 'ethers';
 import { RegistryContract } from 'hooks/Guilds/contracts/useContractRegistry';
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
+import styled from 'styled-components';
 import {
   defaultValues,
   getEditor,
@@ -11,9 +13,19 @@ import {
   DecodedCall,
   SupportedAction,
 } from '../ActionsBuilder/types';
+import { Button } from '../common/Button';
 import { Modal } from '../common/Modal';
 import ContractActionsList from './ContractActionsList';
 import ContractsList from './ContractsList';
+
+export const EditorWrapper = styled.div`
+  margin: 1.25rem;
+`;
+
+export const BlockButton = styled(Button)`
+  margin-top: 1rem;
+  width: 100%;
+`;
 
 interface ActionModalProps {
   isOpen: boolean;
@@ -28,7 +40,12 @@ const ActionModal: React.FC<ActionModalProps> = ({
 }) => {
   const { guild_id: guildId } = useParams<{ guild_id?: string }>();
 
+  // Supported Actions
   const [selectedAction, setSelectedAction] = useState<SupportedAction>(null);
+  const [selectedActionContract, setSelectedActionContract] =
+    useState<utils.Interface>(null);
+
+  // Generic calls
   const [selectedContract, setSelectedContract] =
     useState<RegistryContract>(null);
   const [selectedFunction, setSelectedFunction] = useState<string>(null);
@@ -69,7 +86,12 @@ const ActionModal: React.FC<ActionModalProps> = ({
 
     if (selectedAction) {
       const Editor = getEditor(selectedAction);
-      return <Editor decodedCall={data} updateCall={setData} />;
+      return (
+        <EditorWrapper>
+          <Editor decodedCall={data} updateCall={setData} />
+          <BlockButton onClick={saveSupportedAction}>Save Action</BlockButton>
+        </EditorWrapper>
+      );
     }
 
     return (
@@ -87,6 +109,7 @@ const ActionModal: React.FC<ActionModalProps> = ({
       setSelectedContract(null);
     } else if (selectedAction) {
       setSelectedAction(null);
+      setSelectedActionContract(null);
     }
 
     setData(null);
@@ -100,6 +123,19 @@ const ActionModal: React.FC<ActionModalProps> = ({
     defaultDecodedAction.decodedCall.callType = action;
     setData(defaultDecodedAction.decodedCall);
     setSelectedAction(action);
+    setSelectedActionContract(defaultDecodedAction.contract);
+  }
+
+  function saveSupportedAction() {
+    if (!selectedAction || !data || !setSelectedActionContract) return;
+
+    const decodedAction: DecodedAction = {
+      decodedCall: data,
+      contract: selectedActionContract,
+    };
+
+    onAddAction(decodedAction);
+    setIsOpen(false);
   }
 
   return (
@@ -108,7 +144,7 @@ const ActionModal: React.FC<ActionModalProps> = ({
       onDismiss={() => setIsOpen(false)}
       header={getHeader()}
       maxWidth={300}
-      backnCross={!!selectedContract}
+      backnCross={!!selectedAction || !!selectedContract}
       prevContent={goBack}
     >
       {getContent()}
